thuraya_flask_api/requirements.txt
2captcha_python==1.2.4
beautifulsoup4==4.12.3
colorlog==6.8.2
cryptography==39.0.2
Flask==2.2.3
Flask_JWT_Extended==4.6.0
flask_mail==0.10.0
flask_sqlalchemy==3.0.1
fpdf==1.7.2
itsdangerous==2.1.2
mss==9.0.1
mysql_connector==2.2.9
Pillow==10.0.0
PyAutoGUI==0.9.54
PyPDF2==3.0.1
python-dotenv==1.0.0
Requests==2.28.2
selenium==4.18.1
SQLAlchemy==2.0.4
Werkzeug==2.2.3
